Darkmode addition, contrast & layout improvement
* Added automatic browser driven darkmode selection using @media
(prefers-color-scheme: ...) selectors. Colors and sprite file are set
via variables and the previous theme colors are still defined for older
browsers. e.g.

body {
	background: #c1cdc2;
	background: var(--body);
}

* Reduced page contrast and increased menu / sidebar font weight.

* Improved pagination menu sizing.

* Centered the river of news (RON) pagination menu---an inline <ul>
menu---used on the homepage and category lists--by wrapping it in a div
with the following style:

.table {
        display: table;   /* Allow the centering to work */
        margin: 0 auto;
}

This was suggested in an article on css-tricks.com. The complete link is
in a comment in style.css

* Made the RON pagination menu behavior consistant with the dates paging
menu.