Corrected a repeater offset error (reported by WS0Y)
On Sep 15, 2025, WS0Y reported the the Grandin W0ILO node offset listed on the
RRRA "Radio Systems in the Red River Valley" page differed from that node's
repeaterbook.com listing.

https://rrra.org/radios/index.html

https://www.repeaterbook.com/repeaters/details.php?state_id=38&ID=15

A review of the commit history for this file confirmed WS0Y's report.

The Grandin W0ILO node was correctly listed with a '-' offset until that node
was delisted in commit 01345d8 (May 23, 2023).

This node was relisted with the wrong offset in commit 23ab049 (Aug 10, 2024).